refactor(CodeEditor): drop local primitive, consume frappe-ui/code-editor
CodeEditor/CodePreview now ship from frappe-ui (>=1.0.0-beta.9) under the
frappe-ui/code-editor subpath, so the local copy and its CodeMirror deps,
./CodeEditor export, and generated component types are removed.

Rewire CodeEditorField to the published primitive and fix two silent breaks
from its refined API: readonly -> disabled (sibling-field convention), and
the dropped maxHeight prop -> the --cm-max-height CSS var (restores
expand/collapse). Extract the Frappe-fieldtype -> CodeMirror language map to
fieldtypeToLanguage.ts (FormLayout-specific glue) with its own test.
