+++ ui/components.d.ts
@@ -18,10 +18,7 @@ declare module 'vue' {
     ButtonField: typeof import('./src/components/FormLayout/fields/ButtonField.vue')['default']
     CameraSource: typeof import('./src/components/FileUpload/sources/CameraSource.vue')['default']
     CheckField: typeof import('./src/components/FormLayout/fields/CheckField.vue')['default']
-    CodeEditor: typeof import('./src/components/CodeEditor/CodeEditor.vue')['default']
-    'CodeEditor.story': typeof import('./src/components/CodeEditor/stories/CodeEditor.story.vue')['default']
     CodeEditorField: typeof import('./src/components/FormLayout/fields/CodeEditorField.vue')['default']
-    CodePreview: typeof import('./src/components/CodeEditor/CodePreview.vue')['default']
     DateField: typeof import('./src/components/FormLayout/fields/DateField.vue')['default']
     DatetimeField: typeof import('./src/components/FormLayout/fields/DatetimeField.vue')['default']
     DemoCurrencyField: typeof import('./src/components/FormLayout/stories/DemoCurrencyField.vue')['default']
@@ -76,10 +73,7 @@ declare global {
   const ButtonField: typeof import('./src/components/FormLayout/fields/ButtonField.vue')['default']
   const CameraSource: typeof import('./src/components/FileUpload/sources/CameraSource.vue')['default']
   const CheckField: typeof import('./src/components/FormLayout/fields/CheckField.vue')['default']
-  const CodeEditor: typeof import('./src/components/CodeEditor/CodeEditor.vue')['default']
-  const 'CodeEditor.story': typeof import('./src/components/CodeEditor/stories/CodeEditor.story.vue')['default']
   const CodeEditorField: typeof import('./src/components/FormLayout/fields/CodeEditorField.vue')['default']
-  const CodePreview: typeof import('./src/components/CodeEditor/CodePreview.vue')['default']
   const DateField: typeof import('./src/components/FormLayout/fields/DateField.vue')['default']
   const DatetimeField: typeof import('./src/components/FormLayout/fields/DatetimeField.vue')['default']
   const DemoCurrencyField: typeof import('./src/components/FormLayout/stories/DemoCurrencyField.vue')['default']
