+++ ui/package.json
@@ -17,10 +17,6 @@
       "import": "./src/components/FileUpload/index.ts",
       "types": "./src/components/FileUpload/index.ts"
     },
-    "./CodeEditor": {
-      "import": "./src/components/CodeEditor/index.ts",
-      "types": "./src/components/CodeEditor/index.ts"
-    },
     "./vite": {
       "import": "./vite/index.js"
     },
@@ -39,17 +35,6 @@
   "author": "Frappe Technologies Pvt. Ltd.",
   "license": "MIT",
   "dependencies": {
-    "@codemirror/lang-css": "^6.3.0",
-    "@codemirror/lang-html": "^6.4.9",
-    "@codemirror/lang-javascript": "^6.2.2",
-    "@codemirror/lang-json": "^6.0.1",
-    "@codemirror/lang-markdown": "^6.3.1",
-    "@codemirror/lang-python": "^6.1.6",
-    "@codemirror/lang-sass": "^6.0.2",
-    "@codemirror/lang-sql": "^6.8.0",
-    "@codemirror/lang-xml": "^6.1.0",
-    "@codemirror/lang-yaml": "^6.1.3",
-    "codemirror": "^6.0.1",
     "cropperjs": "^2.1.1",
     "leaflet": "^1.9.4",
     "leaflet-draw": "^1.0.4",
